phase3-modernization/api-python/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0
Faker==21.0.0

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2
mypy==1.7.1

# Type stubs
types-Flask==1.1.6

# Debugging
Flask-DebugToolbar==0.14.1
